divide panel in several files
